WIP include a GraphRAG impl based on BAML
