+++ pyproject.toml
@@ -35,7 +35,9 @@ dependencies = [
     "lxml (>=6.0.0,<7.0.0)",
     "ipython (>=9.4.0,<10.0.0)",
     "nltk (>=3.9.1,<4.0.0)",
-    "bs4 (>=0.0.2,<0.0.3)"
+    "bs4 (>=0.0.2,<0.0.3)",
+    "ollama (>=0.5.1,<0.6.0)",
+    "baml-cli (>=0.1.0,<0.2.0)",
 ]
